refactor(sched): move CompileGraph to sched/legacy_pcm_graph
S2 commit 3/5. Pure mechanical move: the module compile graph becomes
the single live implementation under sched/ until S3 replaces it with
the generalized TaskGraph plus the PCM family. No code changes.
