+++ tests/unit/server/compile_graph_tests.cpp
@@ -2,7 +2,7 @@
 #include <random>
 
 #include "test/test.h"
-#include "server/compiler/compile_graph.h"
+#include "sched/legacy_pcm_graph.h"
 
 #include "llvm/ADT/DenseSet.h"
