+++ tests/unit/server/invalidator_tests.cpp
@@ -1,7 +1,7 @@
 #include "test/cdb_helper.h"
 #include "test/temp_dir.h"
 #include "test/test.h"
-#include "server/compiler/compile_graph.h"
+#include "sched/legacy_pcm_graph.h"
 #include "server/compiler/context_resolver.h"
 #include "server/state/invalidator.h"
